+++ src/command_modules/azure-cli-acs/azure/cli/command_modules/acs/_client_factory.py
@@ -26,6 +26,7 @@ def cf_resource_groups(cli_ctx, subscription_id=None):
 
 def get_auth_management_client(cli_ctx, scope=None, **_):
     import re
+    from azure.cli.core.profiles import ResourceType
     from azure.mgmt.authorization import AuthorizationManagementClient
 
     subscription_id = None
@@ -33,8 +34,7 @@ def get_auth_management_client(cli_ctx, scope=None, **_):
         matched = re.match('/subscriptions/(?P<subscription>[^/]*)/', scope)
         if matched:
             subscription_id = matched.groupdict()['subscription']
-    return get_mgmt_service_client(cli_ctx, AuthorizationManagementClient,
-                                   subscription_id=subscription_id)
+    return get_mgmt_service_client(cli_ctx, ResourceType.MGMT_AUTHORIZATION, subscription_id=subscription_id)
 
 
 def get_container_service_client(cli_ctx, **_):